refactored temp into home and removed unused components
