--- jsuite-frontend/src/components/SharedTemp/SharedTemp.styles.ts
@@ -1,11 +1,0 @@
-import styled from "styled-components";
-import { Button } from "react-bootstrap";
-
-export const Title = styled.h1`
-  text-align: center;
-  color: palevioletred;
-`;
-
-export const BootstrapButton = styled(Button)`
-  background: palevioletred;
-`;
