--- jsuite-frontend/src/components/SharedTemp/SharedTemp.tsx
@@ -1,12 +1,0 @@
-import { Title, BootstrapButton } from './SharedTemp.styles'
-
-const SharedTemp = () => {
-    return (
-        <div>
-          <Title>Shared Temp Component</Title>
-          <BootstrapButton id="shared-temp">Bootstrap button :eyes:</BootstrapButton>
-        </div>
-      );
-}
-
-export default SharedTemp;
