+++ jsuite-frontend/src/routes/Home/Home.tsx
@@ -1,17 +1,14 @@
-import {Global, Title} from './Temp.styles'
-import {Outlet, Link} from 'react-router-dom';
-import SharedTemp from '../../components/SharedTemp/SharedTemp';
+import {Global, Title} from './Home.styles'
 import NavBar from '../../components/NavBar/NavBar';
 
-const Temp = () => {
+const Home = () => {
     return (
         <Global>
           <NavBar></NavBar>
           <Title>Welcome to JSuite!</Title>
           {/* Here goes the app's components */}
-          <SharedTemp></SharedTemp>
         </Global>
       );
 }
 
-export default Temp;
+export default Home;
